+++ CopilotKit/examples/next-openai/CHANGELOG.md
@@ -1,10 +1,49 @@
 # web
 
+## 1.4.1-pre.5
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @copilotkit/runtime-client-gql@1.4.1-pre.5
+  - @copilotkit/react-textarea@1.4.1-pre.5
+  - @copilotkit/react-core@1.4.1-pre.5
+  - @copilotkit/react-ui@1.4.1-pre.5
+  - @copilotkit/runtime@1.4.1-pre.5
+  - @copilotkit/shared@1.4.1-pre.5
+
+## 1.4.1-pre.4
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @copilotkit/runtime-client-gql@1.4.1-pre.4
+  - @copilotkit/react-textarea@1.4.1-pre.4
+  - @copilotkit/react-core@1.4.1-pre.4
+  - @copilotkit/react-ui@1.4.1-pre.4
+  - @copilotkit/runtime@1.4.1-pre.4
+  - @copilotkit/shared@1.4.1-pre.4
+
+## 1.4.1-pre.3
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @copilotkit/runtime-client-gql@1.4.1-pre.3
+  - @copilotkit/react-textarea@1.4.1-pre.3
+  - @copilotkit/react-core@1.4.1-pre.3
+  - @copilotkit/react-ui@1.4.1-pre.3
+  - @copilotkit/runtime@1.4.1-pre.3
+  - @copilotkit/shared@1.4.1-pre.3
+
 ## 1.4.1-pre.2
 
 ### Patch Changes
 
-- Add convertActionsToDynamicStructuredTools to sdk-js
+- bump
 - Updated dependencies
   - @copilotkit/runtime-client-gql@1.4.1-pre.2
   - @copilotkit/react-textarea@1.4.1-pre.2
@@ -17,7 +56,7 @@
 
 ### Patch Changes
 
-- add zod conversion
+- bump
 - Updated dependencies
   - @copilotkit/runtime-client-gql@1.4.1-pre.1
   - @copilotkit/react-textarea@1.4.1-pre.1
@@ -30,14 +69,14 @@
 
 ### Patch Changes
 
-- lower case copilotkit property
+- New prerelease
 - Updated dependencies
+  - @copilotkit/runtime-client-gql@1.4.1-pre.0
   - @copilotkit/react-core@1.4.1-pre.0
+  - @copilotkit/runtime@1.4.1-pre.0
+  - @copilotkit/shared@1.4.1-pre.0
   - @copilotkit/react-textarea@1.4.1-pre.0
   - @copilotkit/react-ui@1.4.1-pre.0
-  - @copilotkit/runtime@1.4.1-pre.0
-  - @copilotkit/runtime-client-gql@1.4.1-pre.0
-  - @copilotkit/shared@1.4.1-pre.0
 
 ## 1.4.0