+++ CopilotKit/examples/node-express/CHANGELOG.md
@@ -1,10 +1,37 @@
 # node
 
+## 1.4.1-pre.5
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @copilotkit/runtime@1.4.1-pre.5
+  - @copilotkit/shared@1.4.1-pre.5
+
+## 1.4.1-pre.4
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @copilotkit/runtime@1.4.1-pre.4
+  - @copilotkit/shared@1.4.1-pre.4
+
+## 1.4.1-pre.3
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @copilotkit/runtime@1.4.1-pre.3
+  - @copilotkit/shared@1.4.1-pre.3
+
 ## 1.4.1-pre.2
 
 ### Patch Changes
 
-- Add convertActionsToDynamicStructuredTools to sdk-js
+- bump
 - Updated dependencies
   - @copilotkit/runtime@1.4.1-pre.2
   - @copilotkit/shared@1.4.1-pre.2
@@ -13,7 +40,7 @@
 
 ### Patch Changes
 
-- add zod conversion
+- bump
 - Updated dependencies
   - @copilotkit/runtime@1.4.1-pre.1
   - @copilotkit/shared@1.4.1-pre.1
@@ -22,7 +49,7 @@
 
 ### Patch Changes
 
-- lower case copilotkit property
+- New prerelease
 - Updated dependencies
   - @copilotkit/runtime@1.4.1-pre.0
   - @copilotkit/shared@1.4.1-pre.0