+++ CopilotKit/packages/runtime-client-gql/CHANGELOG.md
@@ -1,30 +1,51 @@
 # @copilotkit/runtime-client-gql
 
+## 1.4.1-pre.5
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @copilotkit/shared@1.4.1-pre.5
+
+## 1.4.1-pre.4
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @copilotkit/shared@1.4.1-pre.4
+
+## 1.4.1-pre.3
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @copilotkit/shared@1.4.1-pre.3
+
 ## 1.4.1-pre.2
 
 ### Patch Changes
 
-- Add convertActionsToDynamicStructuredTools to sdk-js
+- bump
 - Updated dependencies
-  - @copilotkit/runtime@1.4.1-pre.2
   - @copilotkit/shared@1.4.1-pre.2
 
 ## 1.4.1-pre.1
 
 ### Patch Changes
 
-- add zod conversion
+- bump
 - Updated dependencies
-  - @copilotkit/runtime@1.4.1-pre.1
   - @copilotkit/shared@1.4.1-pre.1
 
 ## 1.4.1-pre.0
 
 ### Patch Changes
 
-- lower case copilotkit property
+- New prerelease
 - Updated dependencies
-  - @copilotkit/runtime@1.4.1-pre.0
   - @copilotkit/shared@1.4.1-pre.0
 
 ## 1.4.0